api/requirements.txt
Flask
flask-cors
Flask-SocketIO
google-genai
loguru
moviepy
numpy
pillow
pyttsx3
tqdm
vosk


pydub
pyfiglet
sentence-transformers
opencv-python
moviepy
diffusers 
invisible_watermark
transformers
accelerate
safetensors
serpapi
google-search-results
groq
